Add files via upload
Module 1 final exercise
As a Data Analyst for the Portuguese parks agency, one of your key responsibilities is to assist in wildfire prevention. To support this, you will analyze the Forest Fires dataset, which details 517 fires in Montesinho Natural Park from January 2000 to December 2003. Your goal is to uncover insights into the conditions that contribute to wildfire ignition and the factors influencing the size of the affected areas. Your analysis will play a crucial role in developing strategies to prevent and manage wildfires effectively.