Add files via upload
Module 4 final assignment file
You work for a government agency that needs to renew its vehicle fleet. To be compliant, the agency needs to hit a specific carbon target. Your task is to analyze the emissions data of various car models to ensure compliance with environmental regulations. Your insights will help guide the selection of the most suitable, greener vehicles.